updating configuration
